fix rendering of disc. in report
